Always use the canonicalized absolute pathname when tracking input files
The hash table tracking .gen input files might be confused when the same file
is required via different paths, e.g. #include "foo.gen" and

This should fix the skeleton template sometimes generating duplicate entries
for the list of .gen source files in the component.

This hash table is used when reporting parse errors, but the file name is
already truncated with basename(3), so this change will make no difference. The
table is also used by [dotgen input deps] in the templates (for generating
dependency files for make(1)). Having a realpath(3) there instead of
the raw path should make no difference either.