Define a new genom-IDL type "optional< type >"
The "optional<type>" type definition is meant for efficient serialization of
"type" when the associated data may not always be present. It is semantically
equivalent to a "sequence< type, 1>".

The C mapping is a C structure containing a boolean flag and the associated
data (not memory management involved). The C++ mapping is not finalized yet but
it is currently defined as a template structure similar to the C mapping.

Thanks to adgeroot@ for pointing me to "boost::optional<>" which implements
a similar thing.