Fix one incorrect list vs. string usage in the tcl engine
The tcl 'interp alias' command expects a list of arguments: one must add a {*}
to the result of the [namespace code] call so that it expands properly.

It appears that tcl<8.6 was doing an extra (incorrect) leval of expansion,
thereby hiding this problem, but the bug fix still works for tcl<8.6.