Update codel interface to pass them an additional context pointer
The initial design choice of keeping a minimalistic C codel interface is not
sustainable. The lack of a context pointer requires either global variables or
at least TLS and these are not always available (for instance if several
components must run inside the same thread or process or even as a kernel
module, like in genom3-orocos).

The context that is now passed around is defined as a const pointer on a new
interface, in charge of the exception management. The genom_throw() function
is replaced by a function pointer in this interface, so that there is no
global implementation anymore. Templates are responsible for the actual
function definition, which is completely free.

The affected C functions that now take an additional context pointer are:
 - codels
 - ports and remotes pseudo-objects
 - functions throwing exceptions

As a side effect, the C client interface is also slightly updated to account
for some mistakes that are now exposed. The json string manipulation functions
("print" and "scan") do not need to throw exceptions, so they are changed to
return a simple integer error code, and the function retrieving the last
exception (json_error) must take the genom_client handle so that it can
retrieve the last exception details. Finally, a new function retrieving the
genom context is added (mostly for throwing exceptions or getting the details
of the last thrown one).

The C++ interface is not changed, with the assumption that C++ will
always be able to pass implicit "this" pointers where needed.

This commit breaks all existing components, but the manual changes in the
codels are minimal and will raise proper compile time errors if not done, so
they will be quickly spotted.

Bump version as well as the C client protocol number.