Add a template helper function computing if a codel is thread-safe
This TCL function is not used by genom but provided as a convenience function
for templates.

A codel is considered to be thread-safe if its parameter list does not conflict
with codels running in other tasks. The function considers that there is a
parameter conflict if the codel's input are found in other codels outputs, or
if the codel's output are found in other codels inputs.
The "start" state of permanent activities of tasks are explicitely not
considered, because those are run before any other codel can run.

The analysis is purely static, using the .gen description.
All reachable states are considered, but incompatible activities are accounted
for, as they are not supposed to run concurrently.