Update the FSM behaviour in periodic tasks and add 'pause events'
/!\ This is an incompatible change in the behaviour of periodic tasks.
/!\ Existing components with periodic tasks must be reviewed.

Transitions in an activity FSM are now always executed immediately, no matter
the task context (periodic or not). (while transitions used to happen only once
in a period for periodic tasks).

To relinquish the CPU, FSM can yield to "paused transition" with a statement
like "yield pause::foo" instead of "yield foo" to indicate a transition to the
new state "foo". Codels are supposed to return a comp_pause_foo instead of
a comp_foo in this situation.

A paused transition in the context of a periodic task simply waits until
the next period. In the context a task with no period, it waits until the next
non-paused transition in the component or a successful attribute/function
execution.

The special "sleep" state has disappeared, since the same behaviour is
obtained by using paused transitions.