During the --with-templates processing, do not save a call to genom
In a --with-templates option, it is fragile to save a call to genom if the
output directory has already been generated: the genom tool might have changed
since the last invocation, and things like Makefile.am or configure.ac may have
to be updated. This would be handled to late by the 'regen' rule in the
generated Makefile.am.

genom is smart enough to not touch generated files if they did not change, so
better run it each time the top level configure itself runs.