Implement a tcl procedure computing a list of mutually exclusive codels
In the code generator, [$codel mutex] returns a list of codels or attributes
that do not support concurrent execution with the given $codel.
Similarly, [$service mutex] returns a list of incompatible codels (this can
a non-empty list only if $service is an attribute).

"async codels" are now checked against these functions, and are only allowed
to have any incompatible codels in the start/stop states of permanent
activities.

This replaces the functionality of the "locking.tcl" file, which will be
removed once templates are updated to use these new function.