New definition of the common genom_state port
Activities are new grouped according to the task running them, so that there is
an activity list per task.

The current activity state is dropped: only the fact that an activity is
running is exported. Maintaining the current activity state up-to-date is
typically a noticable performance hit, for no good reason at the moment since
no software is using it.

The task timings are extended to include a 'load' percentage, representing the
time spent in codels vs. the time spent in waiting for the next event. (So this
is not the same load definition as the average number of processes in the
kernel run queue.)

The port definition now includes unbounded sequences, removing any articficial
limit. Fixes #5, #7, #8