Fix the computation of mutually exclusive codels for a validation codel
Services can still run their validation/simple codels in parallel
with a running instance of an incompatible service, because validation is
done before interruption (because if the validation fails, there is no
interruption).

So a validation codel is potentially mutually exclusive with an fsm codel
of an incompatible service.

Spotted by Mohammed Foughali, many thanks!