Implement tcl engine "move-if-change" settings also after interactive merge
It often happens that merging skeleton files with user changes does not lead to
actual change of the merged files. In this case, the "move-if-change" setting
is honored so that destinations files are not touched. This is especially
useful for configure.ac/Makefile.am that trigger a useless lengthy run of
config.status when they are touched but not modified.