Skeleton template generates an optional README.adoc file from the .gen file
By default, skeleton generates an asciidoc documentation of the .gen
interface, that is placed in README.adoc, so that it can be conveniently
displayed in repository browsers that support that (e.g. git.openrobots.org,
github...).

This can be disabled by invoking skeleton with --terse (or just -t).

The default Makefile.am installs the README.adoc and README.html files to
$docdir by using asciidoctor if it is present. The html file is included in the
distribution tarball so that asciidoctor is not a required
dependency. Asciidoctor is used instead of asciidoc/a2x from the original
asciidoc project, since the former seems to have more momentum nowadays, has a
more robust parser and is similar to what https://git.openrobots.org or github
use, so that offline and online outputs are more consistent.