Switch to asciidoctor to generate the documentation
The change is transparent, but allows to drop various quirks that were
necessary with asciidoc python. Also, the default style is cleaner.
The dependencies for doc generation are different (asciidoctor and a couple of
gems), but since the documentation is included in the distribution tarball,
this is also a seamless change.

While here, include the _Sidebar files used by the online navigation menu to
provide the same navigation in the generated static output.