Fix the code generation for returning a simple type by reference
In C, it is invalid to declare a fucntion returning an array type (either
explicitly - syntax error - or via a typedef). However, this is accepted inside
function arguments, where the array simply decays into a pointer.

Thus, the "[argument reference]" engine procedure (that outputs a type
declarator suitable for a function argument expecting a reference) cannot be
used for declaring a function return value. This commit introduces a
"[return-value reference]" procedure that handles the special case of an array
in C.

This problem was triggered when generating the port->data() member function
when a port is declared as an array type, or a fixed-size string (which is a
char array in C).