Fix the declaration of the *_port_json_print() functions in client/c
The *_port_json_print() is always passed a reference to the port data, not a
value. This is because the invoking function is generic and cannot know the
type of the underlying data.

This fixes the binary to json conversion of ports declared as simple
types (as opposed to constructed types which are always passed by reference).

Spotted by Felix Ingrand, thanks!