Avoid over-linking the codels library because of "codels-requires"
In the skeleton .pc codels files, "codels-requires" must not go into the
"Requires" field, as this is a private requirement.

It should rather go in Requires.private, it this would work correctly. But
Requires.private is mis-behaving in that it still wants the required .pc
even without --static and output the cflags as if the requirement was not
private.

As a work-around, the "codels-require" libs are put in a Libs.private field.