Set TCL system encoding to properly handle native paths with NLS characters
The default TCL encoding for native strings is iso8859-1, which is almost never
correct, so one need to set Tcl_SystemEncoding() to a sane value.

Tcl can guess it from the environment with
Tcl_GetEncodingNameFromEnvironment() and this works well in most
cases.

However, genom can be run from autoconf scripts and it happens that autoconf
unconditionnaly resets the NLS to C. This triggers the use of TCL's default of
iso8859-1, which is wrong when paths have non ASCII characters. Since there is
no way to convince autoconf that the user environment is right, nor to convince
Tcl to use another default than iso8859-1, Tcl_GetEncodingNameFromEnvironment()
is not used and a custom routine that falls back to utf-8 is used instead.

Should fix #224