Handle NLS characters in IDL mappings
IDL mappings source code may contain non-ASCII characters in C comments or
source files path names (comments are embedded in the generated sources and
source files path names in generated #line directines).

When using such input strings in a TCL reflected channel, the channel encoding
must be set to 'binary' to avoid duplicate conversion to internal UTF-8 (at
the reflected channel implementation level and in general TCL channel handling
functions). The previous 'identity' encoding that was used has been obsoleted
in Tcl-8.7 and it was misused anyway.

While here, fix an issue in the 'write' implementation of the reflected
channel: it is expected to return the number of characters written, not bytes
(this is not properly documented in Tcl, and this was failing with non-ASCII
characters).

Should fix #224