Put dependent libraries in _LIBADD instead of _LDFLAGS
This is a relatively recent change in automake that apparently makes it more
strictly conform to the documentation: dependent libraries must go to _LIBADD
and not _LDFLAGS. It was working before probably because _LDFLAGS was added at
the end of the link command line. Now it's at the beginning, effectively
ignoring dependencies that are put there.