Add engine function [$component codels ?validate|simple|fsm? ...]
The code generator engine is extended with a [$component codels] procedure.
It returns the list of all codels defined in the tasks and services of a
component. Optionally, the list may be filtered with one or more keyword from
the list "validate", "simple" or "fsm", returning only the codels of that
kind. Multiple filters may be specified, they are combined with an implicit
"or".