Fix gcc-8 -Wparentheses paranoia in generated code
gcc-8 may now complain about unnecessary parentheses in a type declaration.

This may happen when declaring an unbounded IDL sequence of arrays or
strings. In this case, the C sequence buffer must be a pointer on a array type,
which requires parentheses around the pointer declaration. For that reason,
parentheses were always added, since they were harmless in other cases.

Since this is no longer true, one have to guess when those parentheses are
really necessary. This is done with the heuristic that if the declared variable
name starts with a *, parentheses are required ...