Update profiling API to make it component-idependent
In order to deal with multi-component generation, the profiling API is made
global instead of per-component. This means that only one writer thread will
handle the logs of all components.

The component instance name is added to each profiled events, and not passed
to the profiling initialization function.

The default event queue size can be changed at code generation time.

Bump version to 2.99.38.