[profundis] Decrease memory footprint
Make sure to share common strings in events. Since events mostly consist of the
same strings again and again, sharing identical strings drammatically reduce
memory usage.

Also, store events as lists instead of dictionaries: the lists memory footprint
is much smaller, and the dictionary was just used as a keyed list anyway.
This simplifies sorting and filtering of events, as only one list is managed
(instead of a list of dicts and a list of timestamps).

Bump version.