Enhance the template [mappings] procedure to let it generate codel signatures
The language::mappings now accepts a list of objects that can be types or
codels. For codels, the generated mapping is the signature of the codel, in the
current programming language.

This allows to factor out the generation of the #include lines for the
required external headers (stdint.h and the like) and to address #270.

Bump version.